# go-backend/internal/api/boilerplate-types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package api

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Auth defines model for Auth.
type Auth struct {
	// Jwt The user's JWT
	Jwt *string `json:"jwt,omitempty"`

	// Username The user's username
	Username *string `json:"username,omitempty"`
}

// Error defines model for Error.
type Error struct {
	Code    *int32  `json:"code,omitempty"`
	Message *string `json:"message,omitempty"`
}

// User defines model for User.
type User struct {
	// Email The user's email
	Email *string             `gorm:"unique" json:"email,omitempty"`
	Id    *openapi_types.UUID `gorm:"primaryKey" json:"id,omitempty"`

	// Password The user's password
	Password *string `json:"password,omitempty"`

	// Status The user's status
	Status *string `json:"status,omitempty"`

	// Username The user's name
	Username *string `gorm:"unique" json:"username,omitempty"`
}

// UserCollection defines model for UserCollection.
type UserCollection = []User

// AuthenticateJSONRequestBody defines body for Authenticate for application/json ContentType.
type AuthenticateJSONRequestBody = User

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody = User

// UpdateUserJSONRequestBody defines body for UpdateUser for application/json ContentType.
type UpdateUserJSONRequestBody = User